Preserve useful changes mixed in with controlled list work
