+++ arches/app/src/arches/types.ts
@@ -1,0 +1,8 @@
+export type Language = {
+    code: string;
+    default_direction: "ltr" | "rtl";
+    id: number;
+    isdefault: boolean;
+    name: string;
+    scope: string;
+};
